Remove old klage logic.
