+++ src/components/snarveier/Snarveier.tsx
@@ -3,21 +3,17 @@ import { getTranslations } from "next-intl/server";
 
 import { hentAlleSaker } from "../../generated/ssr/saks-oversikt-controller/saks-oversikt-controller";
 import { hentUtbetalinger } from "../../generated/ssr/utbetalinger-controller/utbetalinger-controller";
-import { getFlag, getToggles } from "../../featuretoggles/unleash";
 
 import SoknaderSnarvei from "./SoknaderSnarvei";
 import UtbetalingerSnarvei from "./UtbetalingerSnarvei";
-import KlagerSnarvei from "./KlagerSnarvei";
 
 const Snarveier = async () => {
-    const [alleSakerResponse, utbetalingerResponse, toggles, t] = await Promise.all([
+    const [alleSakerResponse, utbetalingerResponse, t] = await Promise.all([
         hentAlleSaker(),
         hentUtbetalinger(),
-        getToggles(),
         getTranslations("Snarveier"),
     ]);
-    const klageFlag = getFlag("sosialhjelp.innsyn.klage_enabled", toggles);
-    if (!utbetalingerResponse.length && !alleSakerResponse.length && !klageFlag.enabled) {
+    if (!utbetalingerResponse.length && !alleSakerResponse.length) {
         return null;
     }
     return (
@@ -27,7 +23,6 @@ const Snarveier = async () => {
             </Heading>
             {alleSakerResponse.length > 0 && <SoknaderSnarvei />}
             {utbetalingerResponse.length > 0 && <UtbetalingerSnarvei />}
-            {klageFlag.enabled && <KlagerSnarvei />}
         </VStack>
     );
 };
